backend/json_api/requirements.txt
Flask==2.0.1
APScheduler==3.7.0
beautifulsoup4==4.9.3
requests==2.26.0
Flask-CORS==3.0.10
azure-storage-blob
pytz

